Fix celery worker and enable back improved tests
