+++ website/example_config.py
@@ -81,7 +81,7 @@ USE_CELERY = False
 CELERY_BROKER_URL = 'amqp://guest@localhost//'
 CELERY_RESULT_BACKEND = 'redis://'
 CELERY_TASK_SERIALIZER = 'auth'
-CELERY_RESULT_SERIALIZER = 'pickle'
+CELERY_RESULT_SERIALIZER = 'auth'
 CELERY_ACCEPT_CONTENT = ['auth']
 CELERY_IGNORE_RESULT = False
 CELERY_SECURITY_KEY = '../celery/worker.key'
