+++ website/views/search.py
@@ -1,7 +1,5 @@
-import codecs
 import pickle
 from collections import defaultdict
-from typing import Dict
 from urllib.parse import unquote
 
 from flask import make_response, redirect, abort
@@ -24,21 +22,20 @@ from models import (
     OrderedDict,
     List,
 )
+from search.filters import SearchViewFilters
 from search.mutation import MutationSearch
 from models import Gene
 from models import Mutation
 from models import UsersMutationsDataset
 from sqlalchemy import exists, or_, text
-from helpers.filters.manager import quote_if_needed, FilterManager
-from helpers.filters import Filter
+from helpers.filters.manager import quote_if_needed
 from helpers.widgets import FilterWidget
 from search.mutation_result import SearchResult
+from search.task import SearchTask, search_task
 from views.gene import prepare_subqueries
 from search.protein_mutations import get_protein_muts
 from database import db, levenshtein_sorted, bdb
 from search.gene import GeneMatch, search_feature_engines
-
-search_features = [engine.name for engine in search_feature_engines]
 
 
 def create_engines(options=None):
@@ -122,55 +119,6 @@ def search_proteins(
     return results[:limit]
 
 
-class Feature:
-    """Target class for feature filtering"""
-    pass
-
-
-class Search:
-    pass
-
-
-class SearchViewFilters(FilterManager):
-
-    def __init__(self, **kwargs):
-
-        available_features = search_features
-        active_features = set(available_features) - {'summary'}
-
-        filters = [
-            # Why default = False? Due to used widget: checkbox.
-            # It is not possible to distinguish between user not asking for
-            # all mutations (so sending nothing in post, since un-checking it
-            # will cause it to be skipped in the form) or user doing nothing
-
-            # Why or? Take a look on table:
-            # is_ptm    show all muts (by default only ptm)     include?
-            # 0         0                                       0
-            # 0         1                                       1
-            # 1         0                                       1
-            # 1         1                                       1
-            Filter(
-                Mutation, 'is_ptm', comparators=['or'],
-                default=False
-            ),
-            Filter(
-                Protein, 'has_ptm_mutations', comparators=['eq'],
-                as_sqlalchemy=True
-            ),
-            Filter(
-                Feature, 'name', comparators=['in'],
-                default=list(active_features),
-                choices=available_features,
-            ),
-            Filter(
-                Search, 'query', comparators=['eq'],
-            ),
-        ]
-        super().__init__(filters)
-        self.update_from_request(request)
-
-
 def make_widgets(filter_manager):
     return {
         'proteins': {
@@ -200,39 +148,6 @@ def make_widgets(filter_manager):
             ),
         ]
     }
-
-
-class SearchTask:
-
-    def __init__(self, vcf_file, textarea_query: str, filter_manager: SearchViewFilters, dataset_uri=None):
-        self.vcf_file = vcf_file
-        self.textarea_query = textarea_query
-        self.filter_manager = filter_manager
-        self.dataset_uri = dataset_uri
-
-    def serialize(self) -> Dict:
-        return {
-            'vcf_file': self.vcf_file,
-            'textarea_query': self.textarea_query,
-            'filter_manager': codecs.encode(pickle.dumps(self.filter_manager), 'base64').decode(),
-            'dataset_uri': self.dataset_uri
-        }
-
-    @classmethod
-    def from_serialized(cls, vcf_file, textarea_query, filter_manager, dataset_uri):
-        return cls(
-            vcf_file,
-            textarea_query,
-            pickle.loads(codecs.decode(filter_manager.encode(), 'base64')),
-            dataset_uri
-        )
-
-
-@celery.task
-def search_task(task_data):
-    task = SearchTask.from_serialized(**task_data)
-    mutation_search = MutationSearch(task.vcf_file, task.textarea_query, task.filter_manager)
-    return mutation_search, task.dataset_uri
 
 
 class SearchView(FlaskView):
@@ -407,6 +322,9 @@ class SearchView(FlaskView):
                 )
                 return redirect(url_for('SearchView:mutations'))
             mutation_search, dataset_uri = celery_task.result
+
+            from helpers.pickle import unpickle_str
+            mutation_search = unpickle_str(mutation_search)
 
             if dataset_uri:
                 url = url_for(
